+++ ai_proposal_system/ai_proposal_system/admin-client/src/api/users.js
@@ -1,7 +1,7 @@
 import api from './axios';
 
-export const getUsers = () => api.get('/users');
-export const getUser = (id) => api.get(`/users/${id}`);
-export const createUser = (data) => api.post('/users', data);
-export const updateUser = (id, data) => api.put(`/users/${id}`, data);
-export const deleteUser = (id) => api.delete(`/users/${id}`);
+export const getUsers = () => api.get('/admin/users');
+export const getUser = (id) => api.get(`/admin/users/${id}`);
+export const createUser = (data) => api.post('/admin/users', data);
+export const updateUser = (id, data) => api.put(`/admin/users/${id}`, data);
+export const deleteUser = (id) => api.delete(`/admin/users/${id}`);
